safari-600.7-branch/LayoutTests/fast/dom/htmlallcollection-detached-node-children-expected.txt

This tests verifies that the name getter on element.children works even when element is not in the document

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS element.children['foo'] is undefined.
PASS span = createElementWithId('foo'); element.appendChild(span); element.children['foo'] is span
PASS document.all['foo'] is undefined.
PASS document.body.appendChild(span); document.all['foo'] is span
PASS successfullyParsed is true

TEST COMPLETE
